+++ src/Plugin/Events/PluginsBooted.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CraftCms\Cms\Plugin\Events;
+
+/**
+ * @event PluginsBooted The event that is triggered after all plugins have been booted
+ */
+class PluginsBooted {}