+++ yii2-adapter/legacy/services/Plugins.php
@@ -17,8 +17,8 @@ use CraftCms\Cms\Plugin\Events\PluginEnabling;
 use CraftCms\Cms\Plugin\Events\PluginInstalled;
 use CraftCms\Cms\Plugin\Events\PluginInstalling;
 use CraftCms\Cms\Plugin\Events\PluginRegistered;
+use CraftCms\Cms\Plugin\Events\PluginsBooted;
 use CraftCms\Cms\Plugin\Events\PluginSettingsSaved;
-use CraftCms\Cms\Plugin\Events\PluginsLoaded;
 use CraftCms\Cms\Plugin\Events\PluginsLoading;
 use CraftCms\Cms\Plugin\Events\PluginUninstalled;
 use CraftCms\Cms\Plugin\Events\PluginUninstalling;
@@ -510,7 +510,7 @@ class Plugins extends Component
         );
 
         Event::listen(
-            PluginsLoaded::class,
+            PluginsBooted::class,
             fn() => $pluginService->trigger(self::EVENT_AFTER_LOAD_PLUGINS),
         );